# chore(jsonpath): Improve jsonpath package