@meerumschlungen 2025-09-02
> @meerumschlungen Are you still actively working on this? @TwiN Is it realistic to achieve a merge here in the near future? I think it would also help with my problem #1238.

Without having tested this, I think that the reimplementation might solve your problem. I could create a unit test for this in this branch.